requirements.txt
certifi==2018.11.29
chardet==3.0.4
clearbit==0.1.7
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.8
names==0.3.0
PyJWT==1.7.1
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
